requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py.git@master
python-dotenv
youtube_dl
gtts
PyNaCl
# discord-py-interactions
# tenacity
watchgod
replicate
numpy
